authenticator/requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.1.1
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
sqlparse==0.3.0
